# gateway/docs/openapi.yaml
openapi: 3.0.3
info:
  title: VetManager Pro API
  description: |
    API REST para el sistema veterinario SaaS multinacional VetManager Pro.

    **Autenticación:** Bearer JWT (access token de 15min + refresh token de 7 días)

    **Multi-tenant:** Todos los endpoints filtran datos por `branchId` del token JWT.

    **RBAC:** Cada endpoint requiere un permiso específico según el rol del usuario.
  version: "1.0.0"
  contact:
    name: VetManager Pro
    email: api@vetmanagerpro.com
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.vetmanagerpro.com/api/v1
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /patients:
    get:
      tags: [Patients]
      summary: Listar pacientes
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: speciesId
          schema: { type: integer }
        - in: query
          name: species
          schema: { type: string }
        - in: query
          name: isActive
          schema: { type: string, enum: [true, false, all], default: 'true' }
        - in: query
          name: is_active
          schema: { oneOf: [{ type: string, enum: [true, false, all] }, { type: integer, enum: [0, 1] }] }
        - in: query
          name: page
          schema: { type: string, pattern: '^[0-9]+$' }
        - in: query
          name: limit
          schema: { type: string, pattern: '^[0-9]+$' }
      responses:
        '200': { description: Lista paginada de pacientes }
    post:
      tags: [Patients]
      summary: Crear nuevo paciente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, speciesId, sex, primaryOwnerId]
              properties:
                name:           { type: string }
                speciesId:      { type: integer }
                breedId:        { type: integer }
                sex:            { type: string, enum: [male, female, unknown] }
                birthdate:      { type: string, format: date }
                chipNumber:     { type: string }
                weightKg:       { type: number }
                primaryOwnerId: { type: integer }
      responses:
        '201': { description: Paciente creado }
...